feat(game): 实现技能系统与战斗流程
- 修改属性类型为Long以支持更大数值范围
- 添加技能相关类和工厂方法
- 实现基础战斗循环逻辑
- 添加玩家与Boss的回合制战斗处理
- 实现行动顺序控制机制- 更新装备属性计算方式
- 完善游戏实体属性初始化逻辑

+++ src/main/kotlin/cn/chahuyun/omr/game/PlayerUserExpand.kt
@@ -133,10 +133,10 @@ val PlayerUser.ccritDamage: Int
  *
  * @return 额外速度值，默认为0
  */
-val PlayerUser.espeed: Long
+val PlayerUser.espeed: Int
     get() = equipmentColumn.allEquipment().sumOf {
         it.propertyList.filter { at -> at.type == PropertyType.SPEED }.sumOf { et -> et.value }
-    }.toLong()
+    }
 
 /**
  * 计算玩家用户的总速度
@@ -144,7 +144,7 @@ val PlayerUser.espeed: Long
  *
  * @return 玩家用户的总速度
  */
-val PlayerUser.cspeed: Long
+val PlayerUser.cspeed: Int
     get() = this.speed + this.espeed
